Rename .jsx to .js
