+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "wiki-table-editor-2",
-  "main": "src/wiki-table-editor.jsx",
+  "main": "src/wiki-table-editor.js",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
